+++ CHANGELOG.md
@@ -20,3 +20,4 @@
 - [feat:support consul config.](https://github.com/Tencent/spring-cloud-tencent/pull/1352)
 - [feat:add trace report support.](https://github.com/Tencent/spring-cloud-tencent/pull/1353)
 - [fix: fix lossless test case unstable issues.](https://github.com/Tencent/spring-cloud-tencent/pull/1355)
+- [feat:support consul service update task.](https://github.com/Tencent/spring-cloud-tencent/pull/1357)